Move keto implementation to a separete package

Update claimMembership function name
