+++ auth/keto/doc.go
@@ -1,0 +1,5 @@
+// Copyright (c) Mainflux
+// SPDX-License-Identifier: Apache-2.0
+
+// Package keto contains PolicyAgent implementation using Keto.
+package keto
